fix: move logic to use case
