+++ src/modules/trash/trash.usecase.spec.ts
@@ -9,7 +9,11 @@ import {
 import { getModelToken } from '@nestjs/sequelize';
 import { User } from '../user/user.domain';
 import { SequelizeUserRepository, UserModel } from '../user/user.repository';
-import { Folder, FolderAttributes } from '../folder/folder.domain';
+import {
+  Folder,
+  FolderAttributes,
+  FolderOptions,
+} from '../folder/folder.domain';
 import { FileUseCases } from '../file/file.usecase';
 import { FolderUseCases } from '../folder/folder.usecase';
 import { ShareUseCases } from '../share/share.usecase';
@@ -88,6 +92,10 @@ describe('Trash Use Cases', () => {
     service = module.get<TrashUseCases>(TrashUseCases);
     fileUseCases = module.get<FileUseCases>(FileUseCases);
     folderUseCases = module.get<FolderUseCases>(FolderUseCases);
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
   });
 
   it('should be defined', () => {
@@ -336,6 +344,200 @@ describe('Trash Use Cases', () => {
       expect(folderUseCases.deleteFolderPermanently).toHaveBeenCalledTimes(
         foldersIdToDelete.length,
       );
+    });
+  });
+  describe('Update Trashed Items Parent Folder', () => {
+    const file1 = File.build({
+      id: 1303581,
+      fileId: '',
+      name: '',
+      type: '',
+      size: null,
+      bucket: '',
+      folderId: 350627,
+      encryptVersion: '',
+      deleted: false,
+      deletedAt: new Date(),
+      userId: 1,
+      modificationTime: new Date(),
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    });
+    const file2 = File.build({
+      id: 11159,
+      fileId: '',
+      name: '',
+      type: '',
+      size: null,
+      bucket: '',
+      folderId: 82815,
+      encryptVersion: '',
+      deleted: true,
+      deletedAt: new Date(),
+      userId: 1,
+      modificationTime: new Date(),
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    });
+    const file3 = File.build({
+      id: 14270,
+      fileId: '',
+      name: '',
+      type: '',
+      size: null,
+      bucket: '',
+      folderId: 578977,
+      encryptVersion: '',
+      deleted: false,
+      deletedAt: new Date(),
+      userId: 1,
+      modificationTime: new Date(),
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    });
+
+    const folderA = Folder.build({
+      id: 90,
+      parentId: 543966,
+      name: 'name',
+      bucket: 'bucket',
+      userId: 1,
+      encryptVersion: '03-aes',
+      deleted: true,
+      deletedAt: new Date(),
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    });
+    const folderB = Folder.build({
+      id: 63491,
+      parentId: 18075,
+      name: 'name',
+      bucket: 'bucket',
+      userId: 1,
+      encryptVersion: '03-aes',
+      deleted: false,
+      deletedAt: new Date(),
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    });
+
+    let getFileByFildeId: jest.SpyInstance<Promise<File>, [fileId: string]>;
+    let getFolder: jest.SpyInstance<Promise<Folder>, [number, FolderOptions?]>;
+    let updateFolderUpdatedAt: jest.SpyInstance<
+      Promise<void>,
+      [folderId: number]
+    >;
+
+    beforeEach(() => {
+      getFileByFildeId = jest.spyOn(fileUseCases, 'getFileByFildeId');
+      getFolder = jest.spyOn(folderUseCases, 'getFolder');
+      updateFolderUpdatedAt = jest.spyOn(
+        folderUseCases,
+        'updateFolderUpdatedAt',
+      );
+    });
+
+    afterEach(() => {
+      jest.resetAllMocks();
+    });
+
+    it('updates all the folders from the files', async () => {
+      const filesIds = [file1, file2, file3].map((file) => file.fileId);
+
+      getFileByFildeId
+        .mockImplementationOnce(() => Promise.resolve(file1))
+        .mockImplementationOnce(() => Promise.resolve(file2))
+        .mockImplementationOnce(() => Promise.resolve(file3));
+
+      await service.updateTrashedItemsParentFolder(filesIds, []);
+
+      expect(getFolder).not.toBeCalled();
+      expect(updateFolderUpdatedAt).toHaveBeenCalledTimes(3);
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(file1.folderId);
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(file2.folderId);
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(file3.folderId);
+    });
+
+    it('updates all the folders parents from folders', async () => {
+      const folderIds = [folderA, folderB].map((folder) => folder.id);
+
+      getFolder
+        .mockImplementationOnce(() => Promise.resolve(folderA))
+        .mockImplementationOnce(() => Promise.resolve(folderB));
+
+      await service.updateTrashedItemsParentFolder([], folderIds);
+
+      expect(getFileByFildeId).not.toBeCalled();
+      expect(updateFolderUpdatedAt).toHaveBeenCalledTimes(2);
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(folderA.parentId);
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(folderB.parentId);
+    });
+
+    it('updates all the folder for files and folders', async () => {
+      const filesIds = [file1, file2, file3].map((file) => file.fileId);
+      const foldersIdS = [folderA, folderB].map((folder) => folder.id);
+
+      getFileByFildeId
+        .mockImplementationOnce(() => Promise.resolve(file1))
+        .mockImplementationOnce(() => Promise.resolve(file2))
+        .mockImplementationOnce(() => Promise.resolve(file3));
+
+      getFolder
+        .mockImplementationOnce(() => Promise.resolve(folderA))
+        .mockImplementationOnce(() => Promise.resolve(folderB));
+
+      await service.updateTrashedItemsParentFolder(filesIds, foldersIdS);
+
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(folderA.parentId);
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(folderB.parentId);
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(file1.folderId);
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(file2.folderId);
+      expect(updateFolderUpdatedAt).toHaveBeenCalledWith(file3.folderId);
+    });
+
+    it('does not fails when a file is not found', async () => {
+      getFileByFildeId.mockImplementationOnce(() => Promise.reject());
+
+      try {
+        await service.updateTrashedItemsParentFolder([file1.fileId], []);
+      } catch (error) {
+        expect(error).not.toBeDefined();
+      }
+
+      expect(updateFolderUpdatedAt).not.toBeCalled();
+    });
+
+    it('does not fails when a folder is not found', async () => {
+      getFolder.mockImplementationOnce(() => Promise.reject());
+
+      try {
+        await service.updateTrashedItemsParentFolder([], [folderA.id]);
+      } catch (error) {
+        expect(error).not.toBeDefined();
+      }
+
+      expect(updateFolderUpdatedAt).not.toBeCalled();
+    });
+
+    it('does not fail if a folder could not be updated', async () => {
+      const filesId = [file3, file2].map((file) => file.fileId);
+      const foldersId = [folderA, folderB].map((folder) => folder.id);
+
+      getFileByFildeId
+        .mockImplementationOnce(() => Promise.resolve(file2))
+        .mockImplementationOnce(() => Promise.resolve(file3));
+
+      getFolder
+        .mockImplementationOnce(() => Promise.resolve(folderA))
+        .mockImplementationOnce(() => Promise.resolve(folderB));
+
+      updateFolderUpdatedAt.mockImplementation(() => Promise.reject());
+
+      try {
+        await service.updateTrashedItemsParentFolder(filesId, foldersId);
+      } catch (error) {
+        expect(error).not.toBeDefined();
+      }
     });
   });
 });
